typos and renaming
